replaced MSVC __int types with default types removed old plugin-sdk types (Int32, UInt32, Float, ...) Replaced Bool32 with bool32 Replaced Int64 with int64 More m_dw,m_w > m_n replacements Fixed some classes/structures CPedCollisionAudio > CAEPedAudioEntity

+++ plugin_sa/game_sa/CCutsceneMgr.h
@@ -104,7 +104,7 @@ public:
     static void DeleteCutsceneData();
     static void DeleteCutsceneData_overlay();
     static void FinishCutscene();
-    static uint64_t GetCutsceneTimeInMilleseconds();
+    static long long GetCutsceneTimeInMilleseconds();
     static bool HasCutsceneFinished();
     static void HideRequestedObjects();
     static void Initialise();
